Now checking report length before processing findings in every module

+++ scanwhisp/modules/niktowrapper/niktowrapper.py
@@ -34,7 +34,7 @@ class scanWhispererNiktoWrapper(scanWhispererBase):
         if verbose:
             self.logger.setLevel(logging.DEBUG)
 
-        self.logger.info('Starting {} whisperer'.format(self.CONFIG_SECTION))
+        self.logger.info('\nStarting NiktoWrapper whisperer')
 
         # if the config is available
         if config is not None:
@@ -96,50 +96,61 @@ class scanWhispererNiktoWrapper(scanWhispererBase):
                 if not files_to_process or len(files_to_process) == 0:
                     self.logger.warn('No new scans to process.')
                 else:
-                    self.logger.info('Processing {} new scans'.format(len(files_to_process)))
+                    self.logger.info('Processing {} new reports'.format(len(files_to_process)))
                             
                     # cycle through every scan available
                     for remote_file_name in files_to_process:
                         
-                        self.logger.info('Processing scan {}'.format(remote_file_name))
+                        self.logger.info('Processing report {}'.format(remote_file_name))
 
                         # Download the file from S3     
                         with yaspin(text="Downloading report", color="cyan") as spinner:
                             try:
                                 report_csv = pd.read_csv(io.StringIO(self.niktowrappers3.download_file(remote_file_name)), na_filter=False)
                             except Exception as e:
-                                self.logger.error('niktowrapper findings download failed: {}'.format(e))  
+                                self.logger.error('NiktoWrapper findings download failed: {}'.format(e))  
                                 return
 
                             spinner.ok("✅")
+
+                        # Check if the report contains some findings
+                        if len(report_csv) > 0:
                         
-                        # Iterate over report lines and push it to Elastic Search
-                        with yaspin(text='Creating documents for {} niktowrapper findings.'.format(report_csv.shape[0]), color="cyan") as spinner:
-                            try:
-                                # Iterate over report rows
-                                for index, finding in report_csv.iterrows():
-                                    self.niktowrapperelk.add_to_queue(finding)
-                            except Exception as e:
-                                self.logger.error('niktowrapper document creation failed: {}'.format(e))   
-                                return
+                            # Iterate over report lines and push it to Elastic Search
+                            with yaspin(text='Creating documents from {} NiktoWrapper findings.'.format(report_csv.shape[0]), color="cyan") as spinner:
+                                try:
+                                    # Iterate over report rows
+                                    for index, finding in report_csv.iterrows():
+                                        self.niktowrapperelk.add_to_queue(finding)
+                                except Exception as e:
+                                    self.logger.error('NiktoWrapper document creation failed: {}'.format(e))   
+                                    return
+
+                                spinner.ok("✅")
+                            
+                            # When document queue is ready, push it
+                            with yaspin(text="Pushing documents", color="cyan") as spinner:
+                                try:
+                                    self.niktowrapperelk.push_queue()  
+                                except Exception as e:
+                                    self.logger.error('NiktoWrapper document queue push failed: {}'.format(e))   
+                                    return
+
+                                spinner.ok("✅")
+
+                            # Delete the file from S3
+                            with yaspin(text="Deleting report from S3", color="cyan") as spinner:
+                                try:
+                                    self.niktowrappers3.delete_file(remote_file_name)
+                                    self.logger.debug('File removed from S3: {}'.format(remote_file_name))
+                                except Exception as e:
+                                    self.logger.error('NiktoWrapper report deletion failed: {}'.format(e))  
+                                    return
 
                             spinner.ok("✅")
-                        
-                        # When document queue is ready, push it
-                        with yaspin(text="Pushing documents", color="cyan") as spinner:
-                            try:
-                                self.niktowrapperelk.push_queue()  
-                            except Exception as e:
-                                self.logger.error('niktowrapper document queue push failed: {}'.format(e))   
-                                return
 
-                            spinner.ok("✅")
-
-                        self.logger.debug('{} niktowrapper records whispered to Elastic Search'.format(report_csv.shape[0]))
-
-                        # Delete the file from S3
-                        self.niktowrappers3.delete_file(remote_file_name)
-                        self.logger.debug('File removed from S3: {}'.format(remote_file_name))
+                        else:
+                            self.logger.warn('Report doesn\'t contain any finding')
 
                 # done
                 self.logger.info('All jobs done.')
